refactor(text-to-image): moves "/toSharkUIOutput/GenerateFromTextResponse.ts" to "/toSharkUIOutput/first.ts"
