+++ src/features/TextToImage/Client/SDXL/toSharkUIOutput/first.ts
